build: make gRPC snapshot regen explicit, not env-gated
The build script's grpc::check() is now read-only: it diffs the
freshly generated stub against the committed snapshot but never
writes the source tree. Snapshot refresh moves to a dedicated
binary `refresh_grpc_snapshot` (feature `grpc-codegen`).

Shared generator helpers live in `build_support/grpc/codegen.rs`;
the write path lives in `build_support/grpc/regen.rs`. The build
script only includes `mod.rs`, the binary only includes
`codegen.rs` + `regen.rs`, so the two compile units stay strictly
read-only and write-only respectively. CONTRIBUTING documents the
new invocation.
