+++ crates/thetadatadx/build_support/mod.rs
@@ -21,10 +21,10 @@ pub fn run() -> Result<(), Box<dyn std::error::Error>> {
     // Drift check runs unconditionally: prost-build re-emits the
     // generated stubs into `$OUT_DIR/beta_endpoints.rs`, and we
     // confirm the regenerated bytes match the file already on disk.
-    // Any divergence between `mdds.proto` (or the codegen logic) and
-    // the committed stubs surfaces as a build failure — no env-var
-    // gating means local pre-push catches drift before CI, and CI
-    // catches it whether the contributor remembered the flag or not.
+    // The check is strictly read-only — when the snapshot needs to
+    // change, contributors run the `refresh_grpc_snapshot` binary
+    // (see the module docs in `grpc/mod.rs`). The build script never
+    // writes to the source tree.
     grpc::check()?;
     endpoints::generate_all()?;
     ticks::generate()?;
